fix: resolve clippy items_after_statements and file size cap
- Move type aliases (Row, StageRow) before first statement in
  list_recent_runs and get_stage_timeline to satisfy clippy::items_after_statements
- Extract ChartSkeleton, formatTimestamp, buildDailyEventCounts to
  ActivityPage.utils.tsx; ActivityPage.tsx now 589 lines (< 600 cap)
